requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.7.0
billiard==3.3.0.23
celery==3.1.23
Django==1.8
django-forms-bootstrap==3.0.1
django-timezone-field==1.3
kombu==3.0.35
python-dateutil==2.5.3
pytz==2016.4
redis==2.10.5
six==1.10.0
wheel==0.26.0
